fix call function and endpoint to call
